test: keep unit tests CPU-only and move CUDA test to tests/gpu
Set CUDA_VISIBLE_DEVICES="" in tests/unit/conftest.py so unit tests always run
on CPU even on a GPU machine: device-agnostic tests take their CPU path, and a
test that explicitly needs CUDA now fails loudly (belongs in tests/gpu) instead
of being silently skipped in CI's CPU-only container.

Move test_awq_lite_uncalibrated_linear_keeps_input_quantizer_enabled (the only
such unit test, NVFP4 CUDA-only) to tests/gpu so it actually runs in CI.
